Outside Whitebox Complete
White box the outside area before the white house. Added pieces of cover that are subject to change, as well as enemy placements in the form of cylinders.

Added a golden path which will be used to guide the players to the warehouse if they choose to ignore the enemy encounters. These pillars will be used as light posts.

Empty spaces around will be filled with set dressing such as trees, and the block out walls have been added.